fix(bgui): complete component migration - move View and PageWrapper to src/components
- Moved View component from packages/bgui/View/ to packages/bgui/src/components/View/
- Moved PageWrapper component from packages/bgui/PageWrapper/ to packages/bgui/src/components/PageWrapper/
- Updated exports in packages/bgui/index.ts to reflect new paths
- All bgui components now follow consistent organization structure
