+++ packages/bgui/index.ts
@@ -1,73 +1,72 @@
 // ENTERPRISE-GRADE UI KIT EXPORTS
 // Brain Game UI components built for scale and maintainability
 
-export { PageWrapper } from "./PageWrapper";
+export type { AccordionProps } from "./src/components/Accordion";
 // All Components from src/components (Enterprise folder-per-component structure)
 export { Accordion } from "./src/components/Accordion";
 export { ActionList, ActionListDivider, ActionListItem } from "./src/components/ActionList";
+export type { ActionListItemProps, ActionListProps } from "./src/components/ActionList/types";
+export type { AlertProps, AlertType, AlertVariant } from "./src/components/Alert";
+export { Alert } from "./src/components/Alert";
 export { Avatar } from "./src/components/Avatar";
+export type { AvatarProps } from "./src/components/Avatar/types";
 export { Badge } from "./src/components/Badge";
+export type { BadgeProps } from "./src/components/Badge/types";
+export type { BreadcrumbItemProps, BreadcrumbProps } from "./src/components/Breadcrumb";
+export { Breadcrumb, BreadcrumbItem } from "./src/components/Breadcrumb";
 export { Button } from "./src/components/Button";
+export type { ButtonProps } from "./src/components/Button/types";
 export { Card } from "./src/components/Card";
+export type { CardProps } from "./src/components/Card/types";
 export { Checkbox } from "./src/components/Checkbox";
+export type { CheckboxProps } from "./src/components/Checkbox/types";
 export { Divider } from "./src/components/Divider";
+export type { DividerProps } from "./src/components/Divider/types";
 export { ErrorBoundary } from "./src/components/ErrorBoundary";
 export { Icon } from "./src/components/Icon";
-export { Image } from "./src/components/Image";
-export { Label } from "./src/components/Label";
-export { Link } from "./src/components/Link";
-export { Alert } from "./src/components/Alert";
-export { Breadcrumb, BreadcrumbItem } from "./src/components/Breadcrumb";
-export { Menu, MenuItem } from "./src/components/Menu";
-export { Modal, ModalFooter, ModalHeader } from "./src/components/Modal";
-export { ProgressBar } from "./src/components/ProgressBar";
-export { RadioGroup } from "./src/components/RadioGroup";
-export { Select, SelectItem } from "./src/components/Select";
-export { Slider } from "./src/components/Slider";
-export { Spinner } from "./src/components/Spinner";
-export { Switch } from "./src/components/Switch";
-export { Tabs } from "./src/components/Tabs";
-export { Text } from "./src/components/Text";
-export { TextInput } from "./src/components/TextInput";
-export { Toast } from "./src/components/Toast";
-export { Tooltip } from "./src/components/Tooltip";
-// Core Layout Components
-export { View } from "./View";
-
-export type { PageWrapperProps } from "./PageWrapper/types";
-export type { AccordionProps } from "./src/components/Accordion";
-export type { ActionListItemProps, ActionListProps } from "./src/components/ActionList/types";
-export type { AvatarProps } from "./src/components/Avatar/types";
-export type { BadgeProps } from "./src/components/Badge/types";
-export type { ButtonProps } from "./src/components/Button/types";
-export type { CardProps } from "./src/components/Card/types";
-export type { CheckboxProps } from "./src/components/Checkbox/types";
-export type { DividerProps } from "./src/components/Divider/types";
 // ErrorBoundaryProps is defined in the component file, not exported separately
 export type { IconProps } from "./src/components/Icon/types";
+export { Image } from "./src/components/Image";
 export type { ImageProps } from "./src/components/Image/types";
+export { Label } from "./src/components/Label";
 export type { LabelProps } from "./src/components/Label/types";
+export { Link } from "./src/components/Link";
 export type { LinkProps } from "./src/components/Link/types";
-export type { AlertProps, AlertType, AlertVariant } from "./src/components/Alert";
-export type { BreadcrumbProps, BreadcrumbItemProps } from "./src/components/Breadcrumb";
+export { Menu, MenuItem } from "./src/components/Menu";
 export type { MenuItemProps, MenuProps } from "./src/components/Menu/types";
+export { Modal, ModalFooter, ModalHeader } from "./src/components/Modal";
 export type { ModalProps } from "./src/components/Modal/types";
+export { PageWrapper } from "./src/components/PageWrapper";
+export type { PageWrapperProps } from "./src/components/PageWrapper/types";
+export { ProgressBar } from "./src/components/ProgressBar";
 export type { ProgressBarProps } from "./src/components/ProgressBar/types";
+export { RadioGroup } from "./src/components/RadioGroup";
 export type { RadioGroupItemProps, RadioGroupProps } from "./src/components/RadioGroup/types";
+export { Select, SelectItem } from "./src/components/Select";
 export type { SelectItemProps, SelectProps } from "./src/components/Select/types";
+export { Slider } from "./src/components/Slider";
 export type { SliderProps } from "./src/components/Slider/types";
+export { Spinner } from "./src/components/Spinner";
 export type { SpinnerProps } from "./src/components/Spinner/types";
+export { Switch } from "./src/components/Switch";
 export type { SwitchProps } from "./src/components/Switch/types";
+export { Tabs } from "./src/components/Tabs";
 export type { TabProps, TabsPanelProps, TabsProps } from "./src/components/Tabs/types";
+export { Text } from "./src/components/Text";
 export type { TextProps } from "./src/components/Text/types";
+export { TextInput } from "./src/components/TextInput";
 export type { TextInputProps } from "./src/components/TextInput/types";
+export { Toast } from "./src/components/Toast";
 export type { ToastProps } from "./src/components/Toast/types";
+export { Tooltip } from "./src/components/Tooltip";
 export type { TooltipProps } from "./src/components/Tooltip/types";
+// Core Layout Components
+export { View } from "./src/components/View";
+// Type Exports (Enterprise TypeScript interfaces)
+export type { ViewProps } from "./src/components/View/types";
 // Constants
 export { Fonts } from "./src/constants/Fonts";
 // Hooks
 export { useControlledState, useFocusManagement, useInteractiveState } from "./src/hooks";
 // Utils
 export { withErrorBoundary } from "./src/utils/withErrorBoundary";
-// Type Exports (Enterprise TypeScript interfaces)
-export type { ViewProps } from "./View/types";
