+++ packages/bgui/src/components/PageWrapper/index.tsx
