+++ packages/bgui/src/components/PageWrapper/styles.ts
