+++ packages/bgui/src/components/View/View.tsx
