+++ packages/bgui/src/components/View/index.tsx
