+++ packages/bgui/src/components/View/types.ts
